refactor(api): move get-progression use-case
