+++ api/lib/application/progressions/progression-controller.js
@@ -1,5 +1,5 @@
 import * as ProgressionSerializer from '../../infrastructure/serializers/jsonapi/progression-serializer.js';
-import { usecases } from '../../domain/usecases/index.js';
+import { evaluationUsecases as usecases } from '../../../src/evaluation/domain/usecases/index.js';
 
 const get = function (request) {
   const userId = request.auth.credentials.userId;
